Suppress some %ld versus %d warnings.
There were several warnings in the uptime calculations, and this
minimal change fixed all of them cleanly.

+++ mud/gmain2.c
@@ -26,7 +26,7 @@ void main(argc,argv)     /* The initial routine */
 int argc;
 char *argv[];
     {
-    long r;
+    int/*long anachronism*/ r;
     FILE *a;
     char user[80],b[40],space[400];
     int num;
